feat: Modularize memory gallery UI components and refactor direct open cards
- Extract memory gallery desktop card, mobile card, and edit form into separate widgets
- Replace inline widget builders in MemoryGallerySettingView with new modular widgets
- Remove duplicated code and improve readability in memory gallery feature
- Refactor DirectOpenSettingView to use BeforeOpenCard and AfterOpenCard widgets instead of inline builders
- Clean up unused import in gacha_setting_view.dart
