+++ lib/features/addgift/presentation/views/gacha_setting_view.dart
@@ -1,7 +1,6 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gifo/features/addgift/model/gacha_content.dart';
 import 'package:go_router/go_router.dart';
